# backend/swaggerDoc.yaml
openapi: 3.0.3
info:
  title: Disaster Management API
  version: 1.0.0
servers:
  - url: http://localhost:5555/api
    description: development server
  - url: https://cdrp-1.onrender.com/api
    description: production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      required: [type, description, location]
      properties:
        _id: { type: string }
        type: { type: string, minLength: 4 }
        title: { type: string, minLength: 4 }
        description: { type: string, minLength: 4, maxLength: 300 }
        image:
          {
            type: array,
            items:
              {
                type: object,
                properties:
                  {
                    url: { type: string, format: uri },
                    cloudinaryId: { type: string },
                  },
              },
          }
        location:
          {
            type: object,
            properties:
              {
                type: { type: string, enum: [Point] },
                coordinates: { type: array, items: { type: number } },
                name: { type: string },
              },
          }
        reportedBy: { type: string }
        incident: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /report:
    get:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 5 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/Report" },
                      },
                  },
              },
          }
    post:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                type: { type: string }
                title: { type: string }
                description: { type: string }
                location:
                  {
                    type: object,
                    properties:
                      {
                        type: { type: string },
                        coordinates: { type: array, items: { type: number } },
                      },
                  }
                image:
                  {
                    type: array,
                    items: { type: string, format: binary },
                    maxItems: 10,
                  }
      responses:
        "201":
          {
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Report" } },
              },
          }
    delete:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { content: { application/json: {} } }
...